trouble shooting undefined issue on city

+++ src/main/java/com/hacks/hackathon/city/City.java
@@ -31,8 +31,8 @@ public class City {
     private Long id;
     private String name;
     private Integer population;
-    private Integer xCoord;
-    private Integer yCoord;
+    private float xCoord;
+    private float yCoord;
     private Long fuel;
     @OneToOne
     @Cascade(CascadeType.ALL)
